+++ src/test/java/be/ugent/idlab/knows/mappingweaver/cores/DBTestCore.java
@@ -1,11 +1,5 @@
 package be.ugent.idlab.knows.mappingweaver.cores;
 
-import org.apache.flink.runtime.testutils.MiniClusterResourceConfiguration;
-import org.apache.flink.test.util.MiniClusterWithClientResource;
-import org.apache.ibatis.jdbc.ScriptRunner;
-import org.junit.ClassRule;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
@@ -14,18 +8,15 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
+import org.apache.ibatis.jdbc.ScriptRunner;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
 @Testcontainers
 public abstract class DBTestCore extends TestCore {
 
     protected final String USERNAME;
     protected final String PASSWORD;
     protected final String DOCKER_TAG;
-    @ClassRule
-    public static MiniClusterWithClientResource flinkCluster = new MiniClusterWithClientResource(
-            new MiniClusterResourceConfiguration.Builder()
-                    .setNumberSlotsPerTaskManager(2)
-                    .setNumberTaskManagers(1)
-                    .build());
 
     protected DBTestCore(String username, String password, String dockerTag) {
         this.USERNAME = username;