+++ src/test/java/be/ugent/idlab/knows/mappingweaver/spec/rml_kgc/RMLLVFNMLTest.java
@@ -26,6 +26,6 @@ public class RMLLVFNMLTest extends TestCore {
     @ParameterizedTest(name = "Index: {index} Filename: {0}")
     @MethodSource("positive")
     public void positiveTest(String directory) throws Exception {
-        this.positiveTest("src/test/resources/spec/rml_kgc/rml-lv-fnml/", directory);
+        this.positiveTest("src/test/resources/spec/rml_kgc/rml-lv-fnml/", directory, false);
     }
 }